refactor: Put common structs in seperate header
