+++ pet_mk_vii_planner/src/CMakeLists.txt
@@ -7,6 +7,7 @@ target_sources(rrt
     ${PROJECT_SOURCE_DIR}/src/goal.cpp
     ${PROJECT_SOURCE_DIR}/src/graph.cpp
     ${PROJECT_SOURCE_DIR}/src/rrt.cpp
+    ${PROJECT_SOURCE_DIR}/src/rrtDefinitions.cpp
     ${PROJECT_SOURCE_DIR}/src/utility/algorithm.cpp
     ${PROJECT_SOURCE_DIR}/src/utility/interpolation.cpp
 )
